+++ pywbemtools/pywbemcli/__init__.py
@@ -39,6 +39,7 @@ from ._click_extensions import *    # noqa: F403,F401
 from ._association_shrub import *   # noqa: F403,F401
 from ._utils import *               # noqa: F403,F401
 from ._cimvalueformatter import *   # noqa: F403,F401
+from ._display_cimobjects import *   # noqa: F403,F401
 
 
 from .._version import __version__  # noqa: F401
