+++ pywbemtools/pywbemcli/_common.py
@@ -29,16 +29,13 @@ try:
 except ImportError:
     from ordereddict import OrderedDict  # pylint: disable=import-error
 
-from pydicti import odicti
 import six
 import click
 import tabulate
 
 from pywbem import CIMInstanceName, CIMInstance, CIMClass, \
     CIMQualifierDeclaration, CIMProperty, CIMClassName, \
-    cimvalue, ValueMapping
-
-from .config import USE_TERMINAL_WIDTH, DEFAULT_TABLE_WIDTH
+    cimvalue
 
 from ._cimvalueformatter import cimvalue_to_fmtd_string
 
@@ -51,7 +48,6 @@ SUBCMD_HELP_TXT = "COMMAND [ARGS] " + CMD_OPTS_TXT
 
 DEFAULT_MAX_CELL_WIDTH = 100
 
-INT_TYPE_PATTERN = re.compile(r'^[su]int(8|16|32|64)$')
 
 ##############################################################
 #
@@ -263,8 +259,8 @@ def pick_one_from_list(context, options, title):
     Raises:
       ValueError if Ctrl-c input from console.
 
-    TODO: Possible Future This could be replaced by the python pick library
-    that would use curses for the selection process.
+    TODO/Future: Possible Future This could be replaced by the python
+    pick library that would use curses for the selection process.
     """
 
     # If there is only a single choice, return that choice.
@@ -278,9 +274,10 @@ def pick_one_from_list(context, options, title):
     click.echo(title)
     for index, str_ in enumerate(options):
         click.echo('{}: {}'.format(index, str_))
+    max_option = len(options) - 1
     selection = None
     msg = 'Input integer between 0 and {} or Ctrl-C to exit selection' \
-        .format(index)
+        .format(max_option)
 
     # Loop for valid user choice until valid choice made or selection aborted
     # by user
@@ -288,7 +285,7 @@ def pick_one_from_list(context, options, title):
         try:
             selection_txt = click.prompt(msg)
             selection = int(selection_txt)
-            if 0 <= selection <= index:
+            if 0 <= selection <= max_option:
                 if context:
                     context.spinner_start()
                 return options[selection]
@@ -814,33 +811,6 @@ def split_str_w_esc(astring, delimiter, escape='\\'):
     return ret
 
 
-def get_cimtype(objects):
-    """
-    Get the cim_type for any returned cim object.  Normally this is the
-    name of the class name except that the classname return from
-    getclass and enumerate class is just unicode string
-    """
-    # associators and references return tuple
-    if isinstance(objects, list):
-        test_object = objects[0]
-    elif objects:
-        test_object = object
-    else:
-        cim_type = 'unknown'
-        return None
-
-    if isinstance(test_object, tuple):
-        # associator or reference class level return is tuple
-        cim_type = test_object[0].__class__.__name__
-    else:
-        cim_type = test_object.__class__.__name__
-
-    # account for fact the enumerate class name operation returns uniicode.
-    if isinstance(test_object, six.string_types):
-        cim_type = 'CIMClassName'
-    return cim_type
-
-
 def process_invokemethod(context, objectname, methodname, options):
     # pylint: disable=line-too-long
     """
@@ -920,13 +890,6 @@ def process_invokemethod(context, objectname, methodname, options):
             click.echo('{}={}'.format(pname, val[0]))
 
 
-####################################################################
-#
-#  Display of CIM objects.
-#
-####################################################################
-
-
 def sort_cimobjects(cim_objects):
     """
     Sort lists of CIMClass, CIMCLassName, CIMQualifierDecl, CIMInstance or
@@ -975,218 +938,6 @@ def sort_cimobjects(cim_objects):
         raise TypeError('{} cannot be sorted'.format(type(cim_objects[0])))
 
     return [sort_dict[key] for key in sorted(sort_dict.keys())]
-
-
-def display_cim_objects_summary(context, objects, output_format):
-    """
-    Display a summary of the objects received. This displays the
-    count of objects.
-    """
-    context.spinner_stop()
-
-    if objects:
-        cim_type = get_cimtype(objects)
-
-        if output_format_is_table(output_format):
-            rows = [[len(objects), cim_type]]
-            click.echo(format_table(rows, ['Count', 'CIM Type'],
-                                    title='Summary of {} returned'
-                                    .format(cim_type),
-                                    table_format=output_format))
-            return
-        click.echo('{} {}(s) returned'.format(len(objects), cim_type))
-
-    else:
-        click.echo('0 objects returned')
-
-
-def display_cim_objects(context, cim_objects, output_format, summary=False,
-                        sort=False):
-    """
-    Display CIM objects in form determined by input parameters.
-
-    Input is either a list of cim objects or a single object. It may be
-    any of the CIM types.  This is used to display:
-
-      * CIMClass
-
-      * CIMClassName:
-
-      * CIMInstance
-
-      * CIMInstanceName
-
-      * CIMQualifierDeclaration
-
-      * Or list of the above
-
-    This function may override output type choice in cases where the output
-    choice is not available for the object type.  Thus, for example,
-    mof output makes no sense for class names. In that case, the output is
-    the str of the type.
-
-    Parameters:
-
-      context (:class:`ContextObj`):
-        Click context contained in ContextObj object.
-
-      objects (iterable of :class:`~pywbem.CIMInstance`,
-        :class:`~pywbem.CIMInstanceName`, :class:`~pywbem.CIMClass`,
-        :class:`~pywbem.CIMClassName`,
-        or :class:`~pywbem.CIMQualifierDeclaration`):
-        Iterable of zero or more CIM objects to be displayed.
-
-      output_format (:term:`string`):
-        String defining the preferred output format. Must not be None since
-        the correct output_format must have been selected before this call.
-        Note that the output formats allowed may depend on a) whether
-        summary is True, b)the specific type because we do not have a table
-        output format for CIMClass.
-
-      summary (:class:`py:bool`):
-        Boolean that defines whether the data in objects should be displayed
-        or just a summary of the objects (ex. count of number of objects).
-    """
-    # Note: In the docstring above, the line for parameter 'objects' was way too
-    #       long. Since we are not putting it into docmentation, we folded it.
-
-    context.spinner_stop()
-
-    if summary:
-        display_cim_objects_summary(context, cim_objects, output_format)
-        return
-
-    if not cim_objects and context.verbose:
-        click.echo("No objects returned")
-        return
-
-    if sort:
-        cim_objects = sort_cimobjects(cim_objects)
-
-    # default when displaying cim objects is mof
-    assert output_format
-
-    if isinstance(cim_objects, (list, tuple)):
-        # Table format output is processed as a group
-        if output_format_is_table(output_format):
-            _print_objects_as_table(cim_objects, output_format, context=context)
-        else:
-            # Call to display each object
-            for obj in cim_objects:
-                display_cim_objects(context, obj, output_format=output_format)
-        return
-
-    # Display a single item.
-    object_ = cim_objects
-    # This allows passing single objects to the table formatter (i.e. not lists)
-    if output_format_is_table(output_format):
-        _print_objects_as_table([object_], output_format, context=context)
-    elif output_format == 'mof':
-        try:
-            click.echo(object_.tomof())
-        except AttributeError:
-            # insert NL between instance names for readability
-            if isinstance(object_, CIMInstanceName):
-                click.echo("")
-                click.echo(object_)
-            elif isinstance(object_, (CIMClassName, six.string_types)):
-                click.echo(object_)
-            else:
-                raise click.ClickException('output_format {} invalid for {} '
-                                           .format(output_format,
-                                                   type(object_)))
-    elif output_format == 'xml':
-        try:
-            click.echo(object_.tocimxmlstr(indent=4))
-        except AttributeError:
-            # no tocimxmlstr functionality
-            raise click.ClickException('Output Format {} not supported. '
-                                       'Default to\n{!r}'
-                                       .format(output_format, object_))
-    elif output_format == 'repr':
-        try:
-            click.echo(repr(object_))
-        except AttributeError:
-            raise click.ClickException('"repr" display of {!r} failed'
-                                       .format(object_))
-
-    elif output_format == 'txt':
-        try:
-            click.echo(object_)
-        except AttributeError:
-            raise click.ClickException('"txt" display of {!r} failed'
-                                       .format(object_))
-    # elif output_format == 'tree':
-    #    raise click.ClickException('Tree output format not allowed')
-    else:
-        raise click.ClickException('Invalid output format {}'
-                                   .format(output_format))
-
-
-def _print_classes_as_table(classes, table_width, table_format):
-    """
-    TODO: Future extend to display classes as a table, showing the
-    properties for each class. This will display the properties that exist in
-    subclasses. The temp output
-    so we could create the function is to just output as mof
-    """
-    # pylint: disable=unused-argument
-
-    for class_ in classes:
-        click.echo(class_.tomof())
-
-
-def format_keys(obj, max_width):
-    """
-    Format the keys of a dictionary of keybindings as text for display. Formats
-    multiple keybindings on each line within the max_width
-
-    Parameters:
-
-      obj (:class:`pwbem.CIMInstanceName`):
-        Instance name from which keybindings are to be extracted for
-        formatting.
-
-    Returns:
-        :term:`string` containing the keys from the input obj formatted for
-        display at within the defined width.
-    """
-    def get_wbemurikeys(obj):
-        """
-        Create wbem_uri from CIMInstanceName and separate out key component
-        for return.
-        """
-        wbem_uri = obj.to_wbem_uri()
-        wbem_uri_keys = wbem_uri[wbem_uri.find('.'):]
-        wbem_uri_keys = wbem_uri_keys[1:]
-        return wbem_uri_keys
-
-    assert isinstance(obj, CIMInstanceName)
-    # clear the host and namespace
-    myobj = obj.copy()
-    myobj.host = None
-    myobj.namespace = None
-    wbem_uri_keys = get_wbemurikeys(myobj)
-
-    # Too long for width. Fold the keys on multiple lines
-    if len(wbem_uri_keys) > max_width:
-        wbem_uri_keys = ''
-        line_len = 0
-        for key, value in myobj.keybindings.items():
-            one_key_obj = get_wbemurikeys((CIMInstanceName('x', {key: value})))
-            if wbem_uri_keys:
-                if line_len + len(one_key_obj) > max_width:
-                    wbem_uri_keys += '\n{}'.format(one_key_obj)
-                    line_len = 0
-                else:
-                    wbem_uri_keys += ',{}'.format(one_key_obj)
-                    line_len += len(one_key_obj) + 1
-
-            else:  # must put on first line even if too long
-                wbem_uri_keys += one_key_obj
-                line_len = len(one_key_obj) + 1
-
-    return wbem_uri_keys
 
 
 def display_text(text, output_format=None):  # pylint: disable=unused-argument
@@ -1251,300 +1002,57 @@ def shorten_path_str(path, replacements, fullpath):
     return name_str
 
 
-def _print_paths_as_table(objects, table_width, table_format):
-    # pylint: disable=unused-argument
+def format_keys(obj, max_width):
     """
-    Display paths as a table. This include CIMInstanceName, ClassPath,
-    and unicode (the return type for enumerateClasses).
-    """
-    title = None
-    if objects:
-        if isinstance(objects[0], six.string_types):
-            title = 'Classnames:'
-            headers = ['Class Name']
-            rows = [[obj] for obj in objects]
-        elif isinstance(objects[0], CIMClassName):
-            title = 'Classnames'
-            headers = ('host', 'namespace', 'class')
-            rows = [[obj.host, obj.namespace, obj.classname] for obj in objects]
-        elif isinstance(objects[0], CIMInstanceName):
-            title = 'InstanceNames: {}'.format(objects[0].classname)
-            host_hdr = 'host'
-            ns_hdr = 'namespace'
-            class_hdr = 'class'
-            host_hdr_len = len(host_hdr) + 4
-            ns_hdr_len = len(ns_hdr) + 3
-            class_hdr_len = len(class_hdr) + 3
-            headers = (host_hdr, ns_hdr, class_hdr, 'keysbindings')
+    Format the keys of a dictionary of keybindings as text for display. Formats
+    multiple keybindings on each line within the max_width
 
-            host_lens = [len(obj.host) for obj in objects if obj.host]
-            host_max = max(host_lens) if host_lens else host_hdr_len
-            ns_lens = [len(obj.namespace) for obj in objects if obj.namespace]
-            ns_max = max(ns_lens) if ns_lens else ns_hdr_len
-            class_lens = [len(obj.classname) for obj in objects]
-            class_max = max(class_lens) if class_lens else class_hdr_len
+    Parameters:
 
-            max_key_len = (table_width) - (host_max + ns_max + class_max + 3)
-            rows = [[obj.host, obj.namespace, obj.classname,
-                     format_keys(obj, max_key_len)] for obj in objects]
-        else:
-            raise click.ClickException("{0} invalid type ({1})for path display".
-                                       format(objects[0], type(objects[0])))
-
-        click.echo(format_table(rows, headers, title=title,
-                                table_format=table_format))
-
-
-def _print_qual_decls_as_table(qual_decls, table_width, table_format):
-    """
-    Display the elements of qualifier declarations as a table with a
-    row for each qualifier declaration and a column for each of the attributes
-    of the qualifier declaration (name, type, Value, Array, Scopes, Flavors.
-
-    The function displays all of the qualifier declarations in the
-    """
-    rows = []
-    headers = ['Name', 'Type', 'Value', 'Array', 'Scopes', 'Flavors']
-    max_column_width = int((table_width / len(headers)) - 4)
-    for q in qual_decls:
-        scopes = '\n'.join([key for key in q.scopes if q.scopes[key]])
-        flavors = []
-        flavors.append('EnableOverride' if q.overridable else 'DisableOverride')
-        flavors.append('ToSubclass' if q.tosubclass else 'Restricted')
-        if q.translatable:
-            flavors.append('Translatable')
-        if sum([len(i) for i in flavors]) >= max_column_width:
-            sep = "\n"
-        else:
-            sep = ", "
-        flavors = sep.join(flavors)
-
-        row = [q.name, q.type, q.value, q.is_array, scopes, flavors]
-        rows.append(row)
-
-    click.echo(format_table(rows, headers, title='Qualifier Declarations',
-                            table_format=table_format))
-
-
-def _format_instances_as_rows(insts, max_cell_width=DEFAULT_MAX_CELL_WIDTH,
-                              include_classes=False, context=None,
-                              prop_names=None):
-    """
-    Format the list of instances properties into as a list of the property
-    values for each instance( a row of the table) gathered into a list of
-    the rows.
-
-    The prop_names parameter is the list of (originally cased) property names
-    to be output, in the desired output order. It could be determined from
-    the instances, but since it is determined already by the caller, it
-    is passed in as an optimization. For test convenience, None is permitted
-    and causes the properties to again be determined from the instances.
-
-    Include_classes for each instance if True. Sets the classname as the first
-    column.
-
-    max_width if not None folds col entries longer than the defined
-    max_cell_width. If max_width is None, the data length is ignored.
-
-    The property values are formatted similar to MOF output. Properties that
-    have a ValueMap qualifier (effectively, in the creation class of the
-    instance) are shown with both the actual property value and the mapped
-    value in parenthesis.
-
-    NOTE: This is a separate function to allow testing of the table formatting
-    independently of print output.
+      obj (:class:`pwbem.CIMInstanceName`):
+        Instance name from which keybindings are to be extracted for
+        formatting.
 
     Returns:
-        list of strings where each string is a row in the table and each
-        item in a row is a cell entry
+        :term:`string` containing the keys from the input obj formatted for
+        display at within the defined width.
     """
-    # Avoid crash deeper in code if max_cell_width is None.
-    if max_cell_width is None:
-        max_cell_width = DEFAULT_MAX_CELL_WIDTH
-    lines = []
+    def get_wbemurikeys(obj):
+        """
+        Create wbem_uri from CIMInstanceName and separate out key component
+        for return.
+        """
+        wbem_uri = obj.to_wbem_uri()
+        wbem_uri_keys = wbem_uri[wbem_uri.find('.'):]
+        wbem_uri_keys = wbem_uri_keys[1:]
+        return wbem_uri_keys
 
-    if prop_names is None:
-        prop_names = sorted_prop_names(insts)
+    assert isinstance(obj, CIMInstanceName)
+    # clear the host and namespace
+    myobj = obj.copy()
+    myobj.host = None
+    myobj.namespace = None
+    wbem_uri_keys = get_wbemurikeys(myobj)
 
-    # Cache of ValueMapping objects for integer-typed properties.
-    # Key: classname.propertyname, both in lower case.
-    # A value of None indicates the property does not have a value mapping.
-    valuemappings = {}
-
-    for inst in insts:
-        if not isinstance(inst, CIMInstance):
-            raise ValueError('Only accepts CIMInstance; not type {}'
-                             .format(type(inst)))
-
-        # Insert classname as first col if flag set
-        line = [inst.classname] if include_classes else []
-
-        # get value for each property in this object
-        for name in prop_names:
-
-            # Account for possible instances without all properties
-            # Outputs empty  string.  Note that instance with no value
-            # results in same output as not instance name.
-            if name not in inst.properties:
-                val_str = ''
-            else:
-                value = inst.get(name)
-                p = inst.properties[name]
-
-                # Cache value mappings for integer-typed properties
-                if INT_TYPE_PATTERN.match(p.type) and context:
-                    vm_key = '{}.{}'.format(
-                        inst.classname.lower(), name.lower())
-                    try:
-                        valuemapping = valuemappings[vm_key]
-                    except KeyError:
-                        try:
-                            valuemapping = ValueMapping.for_property(
-                                context.conn,
-                                context.conn.default_namespace,
-                                inst.classname,
-                                name)
-                        except ValueError:
-                            # Property does not have a value mapping.
-                            valuemapping = None
-                        valuemappings[vm_key] = valuemapping
+    # Too long for width. Fold the keys on multiple lines
+    if len(wbem_uri_keys) > max_width:
+        wbem_uri_keys = ''
+        line_len = 0
+        for key, value in myobj.keybindings.items():
+            one_key_obj = get_wbemurikeys((CIMInstanceName('x', {key: value})))
+            if wbem_uri_keys:
+                if line_len + len(one_key_obj) > max_width:
+                    wbem_uri_keys += '\n{}'.format(one_key_obj)
+                    line_len = 0
                 else:
-                    valuemapping = None
+                    wbem_uri_keys += ',{}'.format(one_key_obj)
+                    line_len += len(one_key_obj) + 1
 
-                if value is None:
-                    val_str = u''
-                else:
-                    val_str, _ = cimvalue_to_fmtd_string(
-                        p.value, p.type, indent=0, maxline=max_cell_width,
-                        line_pos=0, end_space=0, avoid_splits=False,
-                        valuemapping=valuemapping)
+            else:  # must put on first line even if too long
+                wbem_uri_keys += one_key_obj
+                line_len = len(one_key_obj) + 1
 
-            line.append(val_str)
-        lines.append(line)
-
-    return lines
-
-
-def _print_instances_as_table(insts, table_width, table_format,
-                              include_classes=False, context=None):
-    """
-    Print the properties of the instances defined in insts as a table where
-    each row is an instance and each column is a property value.
-
-    All properties in the instance are included.
-
-    The header line consists of the property names.
-
-    The property values are formatted similar to MOF output. Properties that
-    have a ValueMap qualifier (effectively, in the creation class of the
-    instance) are shown with both the actual property value and the mapped
-    value in parenthesis.
-    """
-
-    if table_width is None:
-        table_width = DEFAULT_TABLE_WIDTH
-
-    for inst in insts:
-        if not isinstance(inst, CIMInstance):
-            raise ValueError('Only CIMInstance display allows table output')
-
-    prop_names = sorted_prop_names(insts)
-
-    # Try to estimate max cell width from number of cols
-    # This allows folding long data.  However it is incomplete in
-    # that we do not fold the property name.  Further, the actual output
-    # width of a column involves the tabulate outputter, output_format
-    # so this is not deterministic.
-    if prop_names:
-        num_cols = len(prop_names)
-        max_cell_width = int(table_width / num_cols) - 2
-    else:
-        max_cell_width = table_width
-
-    header_line = []
-    if include_classes:
-        header_line.append("classname")
-    header_line.extend(prop_names)
-
-    # Fold long property names
-    new_header_line = []
-    for header in header_line:
-        if len(header) > max_cell_width:
-            new_header_line.append(fold_strings(header, max_cell_width))
-        else:
-            new_header_line.append(header)
-
-    rows = _format_instances_as_rows(insts, max_cell_width=max_cell_width,
-                                     include_classes=include_classes,
-                                     context=context, prop_names=prop_names)
-
-    title = 'Instances: {}'.format(insts[0].classname)
-    click.echo(format_table(rows, new_header_line, title=title,
-                            table_format=table_format))
-
-
-def sorted_prop_names(insts):
-    """
-    Return the list of (originally cased) property names that is the superset
-    of all properties in the input instances.
-
-    The returned list has the key properties first, followed by the non-key
-    properties. Each group is sorted case insensitively.
-
-    The key properties are determined from the instance paths, if present.
-    The function tolerates it if only some of the instances have a path,
-    and if instances of subclasses have additional keys.
-    """
-
-    all_props = odicti()  # key: org prop name, value: lower cased prop name
-    key_props = odicti()  # key: org prop name, value: lower cased prop name
-    for inst in insts:
-        inst_props = inst.keys()
-        for pn in inst_props:
-            all_props[pn] = pn.lower()
-        if inst.path:
-            key_prop_names = inst.path.keys()
-            for pn in inst_props:
-                if pn in key_prop_names:
-                    key_props[pn] = pn.lower()
-
-    nonkey_props = odicti()  # key: org prop name, value: lower cased prop name
-    for pn in all_props:
-        if pn not in key_props:
-            nonkey_props[pn] = all_props[pn]
-
-    key_prop_list = sorted(key_props.keys(), key=lambda p: p.lower())
-    nonkey_prop_list = sorted(nonkey_props.keys(), key=lambda p: p.lower())
-    key_prop_list.extend(nonkey_prop_list)
-    return key_prop_list
-
-
-def _print_objects_as_table(objects, output_format, context=None):
-    """
-    Call the method for each type of object to print that object type
-    information as a table.
-
-    Output format is retrieved from context.
-    """
-    if USE_TERMINAL_WIDTH:
-        table_width = click.get_terminal_size()[0]
-    else:
-        table_width = DEFAULT_TABLE_WIDTH
-
-    if objects:
-        if isinstance(objects[0], CIMInstance):
-            _print_instances_as_table(objects, table_width, output_format,
-                                      context=context)
-        elif isinstance(objects[0], CIMClass):
-            _print_classes_as_table(objects, table_width, output_format)
-        elif isinstance(objects[0], CIMQualifierDeclaration):
-            _print_qual_decls_as_table(objects, table_width, output_format)
-        elif isinstance(objects[0], (CIMClassName, CIMInstanceName,
-                                     six.string_types)):
-            _print_paths_as_table(objects, table_width, output_format)
-        else:
-            raise click.ClickException("Cannot print {} as table"
-                                       .format(type(objects[0])))
+    return wbem_uri_keys
 
 
 def hide_empty_columns(headers, rows):
@@ -1553,6 +1061,14 @@ def hide_empty_columns(headers, rows):
     The definiton of an empty row is:
     1. All entries for the column in all rows are None or "" if type string.
     2. All entries for the column in all rows are None if number.
+
+    Parameters:
+      headers (list of :term:`string`)
+        The strings that represent the column titles of an array of rows.
+
+      rows (list of list of TBD):
+        The rows of a table where each row is a list of the items that
+        represent the columns of the row.
 
     Returns new rows and headers
     """
@@ -1577,7 +1093,12 @@ def hide_empty_columns(headers, rows):
             format(row, headers)
     for column in range(len(headers) - 1, -1, -1):
         if column_is_empty(rows, column):
-            del headers[column]
+            if isinstance(headers, tuple):
+                headersl = list(headers)
+                del headersl[column]
+                headers = tuple(headersl)
+            else:
+                del headers[column]
             for row in rows:
                 del row[column]
 
@@ -1585,7 +1106,7 @@ def hide_empty_columns(headers, rows):
 
 
 def format_table(rows, headers, title=None, table_format='simple',
-                 sort_columns=None):
+                 sort_columns=None, hide_empty_cols=None):
     """
     General print table function.  Prints a list of lists in a
     table format where each inner list is a row.
@@ -1619,12 +1140,19 @@ def format_table(rows, headers, title=None, table_format='simple',
         right). Note that entries in each row of the columns to be sorted
         must be of the same type (int, str, etc.) to be sortable.
 
+    hide_empty_cols (:class:`py:bool`):
+        If this flag is True any columns that are completely blank are
+        hiddend and the column header is removed from the headers.
+        Uses the function hide_empty_columns
+
     Returns:
         :term:`string`: Returns the formatted table as a string
 
     Raises:
         click.ClickException if invalid table format string
     """
+    if hide_empty_cols:
+        headers, rows = hide_empty_columns(headers, rows)
     if sort_columns is not None:
         if isinstance(sort_columns, int):
             rows = sorted(rows, key=itemgetter(sort_columns))
